move place of function canonical_gpu_arch to avoid uncessisary import that causes error
