+++ projects/rocprofiler-compute/src/rocprof_compute_base.py
@@ -15,6 +15,7 @@ import config
 from argparser import omniarg_parser
 from rocprof_compute_soc.soc_base import OmniSoC_Base
 from roofline.run_benchmark import run_roofline_benchmark
+from utils.gpu_arch import canonical_config_arch
 from utils.logger import (
     console_debug,
     console_error,
@@ -28,7 +29,6 @@ from utils.logger import (
 from utils.mi_gpu_spec import mi_gpu_specs
 from utils.specs import (
     MachineSpecs,
-    canonical_config_arch,
     generate_machine_specs,
 )
 from utils.utils_common import (
